extract PoiDescription.tsx
