+++ src/components/FeaturePanel/ImageSection/ImageSection.tsx
@@ -6,11 +6,9 @@ import React from 'react';
 import styled from 'styled-components';
 import { useFeatureContext } from '../../utils/FeatureContext';
 import { FeatureImage } from './FeatureImage';
-import Maki from '../../utils/Maki';
-import { hasName } from '../../../helpers/featureLabel';
 import { t } from '../../../services/intl';
 import { SHOW_PROTOTYPE_UI } from '../../../config';
-import { Feature } from '../../../services/types';
+import { PoiDescription } from './PoiDescription';
 
 const StyledIconButton = styled(IconButton)`
   svg {
@@ -20,41 +18,13 @@ const StyledIconButton = styled(IconButton)`
   }
 `;
 
-const PoiType = styled.div`
-  color: #fff;
-  margin: 0 auto 0 15px;
-  font-size: 13px;
-  position: relative;
-  width: 100%;
-  svg {
-    vertical-align: bottom;
-  }
-  span {
-    position: absolute;
-    left: 20px;
-  }
-`;
-
-const getSubclass = ({ layer, osmMeta, properties, schema }: Feature) =>
-  schema?.label ||
-  properties.subclass?.replace(/_/g, ' ') ||
-  (layer && layer.id) || // layer.id specified only when maplibre-gl skeleton displayed
-  osmMeta.type;
-
 export const ImageSection = () => {
   const { feature } = useFeatureContext();
   const { properties } = feature;
 
-  const poiType = hasName(feature)
-    ? getSubclass(feature)
-    : t('featurepanel.no_name');
-
   return (
     <FeatureImage feature={feature} ico={properties.class}>
-      <PoiType>
-        <Maki ico={properties.class} invert middle />
-        <span>{poiType}</span>
-      </PoiType>
+      <PoiDescription />
 
       {SHOW_PROTOTYPE_UI && (
         <>
